+++ front/src/hooks/auth/useGetCurrentUser.ts
@@ -1,12 +1,10 @@
-import { useRecoilValue } from "recoil"
-import { currentUserState } from "@/recoil/atoms"
+import { useRecoilValue } from 'recoil'
+import { currentUserState } from '@/recoil/atoms'
 
 export const useGetCurrentUser = () => {
   const currentUser = useRecoilValue(currentUserState)
-  const isAuthChecking = currentUser === undefined
 
   return {
-    currentUser,
-    isAuthChecking
+    currentUser
   }
 }
